Make AffineTransform use double instead of float, fix inline images parsing bug, improve rectangles comparison

+++ layout/src/main/java/com/itextpdf/layout/renderer/ImageRenderer.java
@@ -312,10 +312,10 @@ public class ImageRenderer extends AbstractRenderer {
         t.translate(xDistance, yDistance);
         t.getMatrix(matrix);
         if (fixedXPosition != null) {
-            fixedXPosition += t.getTranslateX();
+            fixedXPosition += (float)t.getTranslateX();
         }
         if (fixedYPosition != null) {
-            fixedYPosition += t.getTranslateY();
+            fixedYPosition += (float)t.getTranslateY();
         }
     }
 }
